+++ src/cmd/mod.rs
@@ -1,7 +1,9 @@
 //! Command execution and sandboxing.
 
+mod process_lines_actions;
 mod sandbox;
 
+pub use process_lines_actions::ProcessLinesActions;
 pub use sandbox::*;
 
 use crate::native;
@@ -9,6 +11,7 @@ use crate::workspace::Workspace;
 use failure::{Error, Fail};
 use futures::{future, Future, Stream};
 use log::{error, info};
+use process_lines_actions::InnerState;
 use std::convert::AsRef;
 use std::env::consts::EXE_SUFFIX;
 use std::ffi::{OsStr, OsString};
@@ -125,7 +128,7 @@ pub struct Command<'w, 'pl> {
     binary: Binary,
     args: Vec<OsString>,
     env: Vec<(OsString, OsString)>,
-    process_lines: Option<&'pl mut dyn FnMut(&str)>,
+    process_lines: Option<&'pl mut dyn FnMut(&str, &mut ProcessLinesActions)>,
     cd: Option<PathBuf>,
     timeout: Option<Duration>,
     no_output_timeout: Option<Duration>,
@@ -240,7 +243,7 @@ impl<'w, 'pl> Command<'w, 'pl> {
     /// let mut ice = false;
     /// Command::new(&workspace, "cargo")
     ///     .args(&["build", "--all"])
-    ///     .process_lines(&mut |line| {
+    ///     .process_lines(&mut |line, _| {
     ///         if line.contains("internal compiler error") {
     ///             ice = true;
     ///         }
@@ -249,7 +252,7 @@ impl<'w, 'pl> Command<'w, 'pl> {
     /// # Ok(())
     /// # }
     /// ```
-    pub fn process_lines(mut self, f: &'pl mut dyn FnMut(&str)) -> Self {
+    pub fn process_lines(mut self, f: &'pl mut dyn FnMut(&str, &mut ProcessLinesActions)) -> Self {
         self.process_lines = Some(f);
         self
     }
@@ -480,7 +483,7 @@ impl OutputKind {
 
 fn log_command(
     mut cmd: StdCommand,
-    mut process_lines: Option<&mut dyn FnMut(&str)>,
+    mut process_lines: Option<&mut dyn FnMut(&str, &mut ProcessLinesActions)>,
     capture: bool,
     timeout: Option<Duration>,
     no_output_timeout: Option<Duration>,
@@ -512,6 +515,7 @@ fn log_command(
         .map(|line| (OutputKind::Stderr, line));
 
     let start = Instant::now();
+    let mut actions = ProcessLinesActions::new();
 
     let output = stdout
         .select(stderr)
@@ -533,21 +537,31 @@ fn log_command(
                 return future::err(Error::from(CommandError::Timeout(timeout.as_secs())));
             }
 
-            if log_output {
-                info!("[{}] {}", kind.prefix(), line);
+            if let Some(f) = &mut process_lines {
+                f(&line, &mut actions);
             }
-            future::ok((kind, line))
+            // this is done here to avoid duplicating the output line
+            let lines = match actions.take_lines() {
+                InnerState::Removed => Vec::new(),
+                InnerState::Original => vec![line],
+                InnerState::Replaced(new_lines) => new_lines,
+            };
+
+            if log_output {
+                for line in &lines {
+                    info!("[{}] {}", kind.prefix(), line);
+                }
+            }
+
+            future::ok((kind, lines))
         })
         .fold(
             (Vec::new(), Vec::new()),
-            move |mut res, (kind, line)| -> Result<_, Error> {
-                if let Some(f) = &mut process_lines {
-                    f(&line);
-                }
+            move |mut res, (kind, mut lines)| -> Result<_, Error> {
                 if capture {
                     match kind {
-                        OutputKind::Stdout => res.0.push(line),
-                        OutputKind::Stderr => res.1.push(line),
+                        OutputKind::Stdout => res.0.append(&mut lines),
+                        OutputKind::Stderr => res.1.append(&mut lines),
                     }
                 }
                 Ok(res)