+++ src/crates/git.rs
@@ -1,5 +1,5 @@
 use super::CrateTrait;
-use crate::cmd::Command;
+use crate::cmd::{Command, ProcessLinesActions};
 use crate::prepare::PrepareError;
 use crate::Workspace;
 use failure::{Error, ResultExt};
@@ -83,7 +83,7 @@ impl CrateTrait for GitRepo {
         //    fata: credential helper '{path}' told us to quit
         //
         let mut private_repository = false;
-        let mut detect_private_repositories = |line: &str| {
+        let mut detect_private_repositories = |line: &str, _actions: &mut ProcessLinesActions| {
             if line.starts_with("fatal: credential helper") && line.ends_with("told us to quit") {
                 private_repository = true;
             }